# static/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoOnCall",
    "description": "基于 LangChain 的智能oncall运维系统",
    "version": "1.2.1"
  },
  "paths": {
...
    "/api/upload/config": {
      "get": {
        "tags": [
          "文件管理"
        ],
        "summary": "Upload Config",
        "description": "Return upload constraints used by the frontend before selecting a file.",
        "operationId": "upload_config_api_upload_config_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/upload": {
      "post": {
        "tags": [
          "文件管理"
        ],
        "summary": "Upload File",
        "description": "上传文件并自动创建向量索引\n\nArgs:\n    file: 上传的文件\n\nReturns:\n    JSONResponse: 上传结果",
        "operationId": "upload_file_api_upload_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadFileResponse"
                }
              }
            }
          },
          "207": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadFileResponse"
                }
              }
            },
            "description": "Multi-Status"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/index_directory": {
      "post": {
        "tags": [
          "文件管理"
        ],
        "summary": "批量索引目录（运维入口）",
        "description": "面向 make upload 等批处理/运维流程；前端工作台主入口仍使用 /upload。",
        "operationId": "index_directory_api_index_directory_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "directory_path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 1024
                },
                {
                  "type": "null"
                }
              ],
              "title": "Directory Path"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/indexing/reports": {
      "get": {
        "tags": [
          "文件管理"
        ],
        "summary": "Knowledge Indexing Reports",
        "description": "Return persisted loader cleaning quality reports and doc_type aggregates.",
        "operationId": "knowledge_indexing_reports_api_knowledge_indexing_reports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Doc Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeIndexingReportsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_api_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeIndexingReportsResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "title": "KnowledgeIndexingReportsResponse",
        "description": "Knowledge indexing quality report response."
      },
...
      "UploadConfigData": {
        "properties": {
          "allowed_extensions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Extensions"
          },
          "max_file_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max File Size"
          },
          "max_file_size_mb": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max File Size Mb"
          }
        },
        "type": "object",
        "required": [
          "max_file_size",
          "max_file_size_mb"
        ],
        "title": "UploadConfigData",
        "description": "Frontend-visible file upload constraints."
      },
      "UploadConfigResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/UploadConfigData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "UploadConfigResponse",
        "description": "File upload configuration response."
      },
      "UploadFileData": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "size": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Size"
          },
          "overwritten": {
            "type": "boolean",
            "title": "Overwritten",
            "default": false
          },
          "indexing_ready": {
            "type": "boolean",
            "title": "Indexing Ready",
            "default": false
          },
          "indexing": {
            "$ref": "#/components/schemas/UploadIndexingStatus"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "file_path",
          "size",
          "indexing"
        ],
        "title": "UploadFileData",
        "description": "Saved file metadata plus indexing readiness."
      },
      "UploadFileResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "$ref": "#/components/schemas/UploadFileData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "UploadFileResponse",
        "description": "File upload and indexing response."
      },
      "UploadIndexingStatus": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "empty",
              "failed"
            ],
            "title": "Status"
          },
          "chunk_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Chunk Count"
          },
          "duration_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Duration Ms"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "cleaning": {
            "additionalProperties": true,
            "type": "object",
            "title": "Cleaning"
          }
        },
        "type": "object",
        "required": [
          "status",
          "chunk_count",
          "duration_ms"
        ],
        "title": "UploadIndexingStatus",
        "description": "Public indexing outcome returned by the upload API."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}